Out with the car, in with the abyssal void.
After trying to implement a movement system similar to Brain's, we have
no error messages per say, and the code compiles fine. However we have
the black screen of doom. Will continue to work on this, I have a
feeling the problem may lie in setting the position of the camera or the
model.